+++ CHANGELOG.md
@@ -21,6 +21,8 @@ to [Semantic Versioning](https://semver.org).
 
 ### Fixed
 
+- #31 - IllegalAccessError while request document for a misplaced property value.
+
 ### Security
 
 ## [0.17.0] - 2022-11-07